backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
motor==3.6.0
pydantic==2.9.2
python-dotenv==1.0.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
email-validator==2.1.1

# AI/ML - Using OpenAI (user already has working API key)
openai==1.54.0

# Also keep google-generativeai for ADK compatibility layer
google-generativeai==0.3.2

# Web scraping
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3

# Optional: Observability (for production)
# Uncomment if you want structured logging and tracing
# opentelemetry-api==1.20.0
# opentelemetry-sdk==1.20.0
# opentelemetry-exporter-otlp==1.20.0
# structlog==23.1.0

# Optional: Database support for sessions (for production)
# Uncomment if using PostgreSQL for session storage
# psycopg2-binary==2.9.9


